// Laudari.WebApp/laudari-webapp/src/app/core/models/ultrassom/liquidoLivre/liquido-livre.enum.ts
export enum liquidoLibreClassificacao {
    Ausente = 0,
    Abdome = 1,
    Pelve = 2,
    FossaHepatorrenal = 3,
    FlancoDireito = 4,
    FlancoEsquerdo = 5,
    FossaIliacaDireita = 6,
    FossaIliacaEsquerda = 7
}

export enum liquidoLivreQtd{
    PequenaQtd = 0,
    ModeradaQtd = 1,
    GrandeQtd = 2
}